FIX: un-pub hidden test module tw
The tw module contains a reference implementation -- transcribed right
from the paper, using 1-based indexing(!). It's a fun thing, but it was
not intended to be used. Always was hidden from docs and only was public
for benchmarks.

By only compiling the tw module for tests, we have less code to compile
in the crate normally.