Add 'memmem' as a Cargo keyword, remove 'no_std'.
This algorithm is commonly used to implement memmem because it's very
efficient, so let's make it easier to find for those users who don't
know what algorithm they want.

no_std, on the other hand, is already a category so we don't need to
keep it and there should be no more than 5 keywords.