FIX: pcmp pattern and text loads so that they stay in bounds
The simd loads require that we can read all 16 bytes of memory from the
pointer.

Add function mask_load to read from shorter slices of memory if
required. Follow the code, wherever we have the variable "safetext", we
have already ensured we're only reading up until before the last 16
bytes of the input text, so here it is safe to use simd loads.

This way, the code passes fuzz testing without errors, and keeps the
benchmark results.