FIX: Update `pos` in pcmp search loop when we find no match
This is a major bug fix to performance (not correctness).

We forgot to update pos before we leave the main search loop in pcmp
search (and we enter the tail loop). This meant that if there were no
partial matches, we'd often scan the input text twice, and that's once
with the more inefficent tail loop.